requirements.txt
fastapi
uvicorn
google-cloud-asset
pydantic
